Move message classes to common messages.py
