+++ docs/docs/en/api/fastagency/messages/InputResponseModel.md
@@ -8,4 +8,4 @@ search:
   boost: 0.5
 ---
 
-::: fastagency.adapters.nats.messages.InputResponseModel
+::: fastagency.messages.InputResponseModel
